Refactor project structure bench 2680438
