+++ src/evaluation/cache.rs
@@ -1,7 +1,15 @@
-use crate::evaluation::network::{HIDDEN, NETWORK, NUM_BUCKETS};
-use crate::types::bitboard::Bitboard;
-use crate::types::side::Side;
+use crate::board::bitboard::Bitboard;
+use crate::board::side::Side;
+use crate::evaluation::network::{HIDDEN, NETWORK};
+use crate::evaluation::NUM_BUCKETS;
 
+/// Whenever the king changes bucket, a costly full refresh of the accumulator is required. This
+/// service implements a technique to improve the performance of this refresh known as 'Finny tables'.
+///
+/// We keep a cache of the last accumulator and board state used for each bucket. When refreshing,
+/// instead of starting from an empty board, we start from the last board state used for the bucket.
+/// We therefore only need to apply the diff between the last board state and the current board state
+/// to the accumulator.
 #[derive(Clone, Default)]
 #[repr(align(64))]
 pub struct InputBucketCache {
