--- src/evaluation/mod.rs
@@ -1,5 +1,0 @@
-pub mod accumulator;
-pub mod cache;
-pub mod feature;
-pub mod network;
-pub mod simd;
