+++ src/tools/bench.rs
@@ -1,7 +1,7 @@
 use crate::board::Board;
 use crate::search::search;
-use crate::thread::ThreadData;
-use crate::time::SearchLimits;
+use crate::search::thread::ThreadData;
+use crate::search::time::SearchLimits;
 use std::time::Instant;
 
 const BENCH_DEPTH: u64 = 12;
