+++ src/tools/fen.rs
@@ -1,11 +1,12 @@
+use crate::board::castling::Rights;
+use crate::board::file::File;
+use crate::board::piece::Piece;
+use crate::board::rank::Rank;
+use crate::board::side::Side;
+use crate::board::side::Side::{Black, White};
+use crate::board::square::Square;
+use crate::board::zobrist::Zobrist;
 use crate::board::Board;
-use crate::types::castling::Rights;
-use crate::types::piece::Piece;
-use crate::types::side::Side;
-use crate::types::side::Side::{Black, White};
-use crate::types::square::Square;
-use crate::types::{File, Rank};
-use crate::zobrist::Zobrist;
 
 pub const STARTPOS: &str = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
