+++ src/tools/mod.rs
@@ -1,0 +1,7 @@
+pub mod bench;
+pub mod datagen;
+pub mod fen;
+pub mod perft;
+pub mod scharnagl;
+pub mod uci;
+pub mod utils;
