--- src/types/mod.rs
@@ -1,11 +1,0 @@
-pub mod bitboard;
-pub mod square;
-pub mod file;
-pub mod rank;
-pub mod piece;
-pub mod side;
-pub mod castling;
-pub mod ray;
-
-pub use file::File;
-pub use rank::Rank;
